GUI Components
GameBoard is the most important one, it's currently in a very hack mode because I haven't loaded up the other images, but if you run it (obviously in conjunction with the DrawArea and MiscImage classes, it can return your coordinate point (currently set so that each square is 40 x 40)). I will probably edit it so that it can work with a path. 

Images will come later, this is just a temporary commit.