A few minor changes
Tower.java now deletes the Enemy OBJECT from LSD.wave, and it checks for enemies ONLY if the wave is running (waveDone is false) 
LSD.java: a few more debug statements
Enemy.java: Kill counter now increments. 

So: problem. Kill counter should now accurately record kills but around wave 2 or 3, the enemy gets killed (the debug statement says so), but the kill counter does not increment and the next wave is unable to start. The image of the monster is just left there, frozen, and the towers do not shoot at it. Is waveDone broken, or...?